+++ EA test 2/src/controller/playState/entityController/npcController/DocController.java
@@ -16,8 +16,8 @@ public class DocController extends EntityController {
 
 	@Override
 	public void update() {
-		int xDistance = Math.abs(hitbox.x - play.getPlayer().getHitbox().x);
-		int yDistance = Math.abs(hitbox.y - play.getPlayer().getHitbox().y);
+		float xDistance = Math.abs(hitbox.x - play.getPlayer().getHitbox().x);
+		float yDistance = Math.abs(hitbox.y - play.getPlayer().getHitbox().y);
 		
 		if(xDistance < play.getController().getTileSize()*1.5 && yDistance < play.getController().getTileSize()*1.5) {			
 			play.getController().getView().getPlay().getUI().setMessage("premi E per parlare");
